+++ source/Surface.h
@@ -18,7 +18,7 @@ namespace Surface
 	class Sphere : public Base
 	{
 	public:
-		Sphere(const glm::vec3& origin, float radius)
+		Sphere(const glm::dvec3& origin, double radius)
 			: origin(origin), radius(radius) { }
 
 		~Sphere() {};
@@ -26,29 +26,29 @@ namespace Surface
 		virtual bool intersect(const Ray& ray, Intersection& intersection) const;
 
 	private:
-		glm::vec3 origin;
-		float radius;
+		glm::dvec3 origin;
+		double radius;
 	};
 
 	class Triangle : public Base
 	{
 	public:
-		Triangle(const glm::vec3& v0, const glm::vec3& v1, const glm::vec3& v2)
+		Triangle(const glm::dvec3& v0, const glm::dvec3& v1, const glm::dvec3& v2)
 			: v0(v0), v1(v1), v2(v2), E1(v1 - v0), E2(v2 - v0), normal(glm::normalize(glm::cross(E1, E2))) { }
 
 		~Triangle() {};
 
 		virtual bool intersect(const Ray& ray, Intersection& intersection) const;
 
-		glm::vec3 operator()(float u, float v) const
+		glm::dvec3 operator()(double u, double v) const
 		{
 			return (1 - u - v) * v0 + u * v1 + v * v2;
 		}
 
 	private:
-		glm::vec3 v0, v1, v2;
+		glm::dvec3 v0, v1, v2;
 
 		// Pre-computed edges and normal
-		glm::vec3 E1, E2, normal;
+		glm::dvec3 E1, E2, normal;
 	};
 }
